big refactor
* moves things to tools and modules folder
* adds comments for direct invocation
* modifies ols to support new folder structure
* changes to ignore tmptest* files for experimentation
* adds kubernetes python module and sorts requirements
