+++ modules/task_rephraser.py
@@ -1,24 +1,26 @@
 import logging
 import sys
-from model_context import get_watsonx_predictor
+from modules.model_context import get_watsonx_predictor
+from modules.yes_no_classifier import YesNoClassifier
+from modules.task_performer import TaskPerformer
 from langchain.chains import LLMChain
 from langchain.prompts import PromptTemplate
 
 from string import Template
 
-DEFAULT_MODEL = "ibm/granite-13b-instruct-v1"
+DEFAULT_MODEL = "ibm/granite-13b-chat-grounded-v01"
 
 
-class QuestionValidator:
+class TaskRephraser:
     def __init__(self):
         logging.basicConfig(
             stream=sys.stdout,
             format="%(asctime)s [%(name)s] %(levelname)s: %(message)s",
             level=logging.INFO,
         )
-        self.logger = logging.getLogger("question_validator")
+        self.logger = logging.getLogger("task_rephraser")
 
-    def validate_question(self, conversation, query, **kwargs):
+    def rephrase_task(self, conversation, task, original_query, **kwargs):
         if "model" in kwargs:
             model = kwargs["model"]
         else:
@@ -34,7 +36,7 @@ class QuestionValidator:
 
         # TODO: must be a smarter way to do this
         settings_string = Template(
-            '{"conversation": "$conversation", "query": "$query","model": "$model", "verbose": "$verbose"}'
+            '{"conversation": "$conversation", "task": "$task", "query": "$original_query","model": "$model", "verbose": "$verbose"}'
         )
 
         self.logger.info(
@@ -42,7 +44,8 @@ class QuestionValidator:
             + " call settings: "
             + settings_string.substitute(
                 conversation=conversation,
-                query=query,
+                task=task,
+                original_query=original_query,
                 model=model,
                 verbose=verbose,
             )
@@ -51,54 +54,42 @@ class QuestionValidator:
         prompt_instructions = PromptTemplate.from_template(
             """
 Instructions:
-- You are a question classifying tool
-- You are an expert in kubernetes and openshift
-- Your job is to determine if a question is about kubernetes or openshift and to provide a one word response
-- If a question is not about kubernetes or openshift, answer with only the word "INVALID"
-- If a question is about kubernetes or openshift, answer with only the word "VALID"
-- Do not provide explanation, only respond with the single chosen word
+- You are a helpful assistant.
+- Your job is to combine the information from the task and query into a single, new task.
+- Base your answer on the provided task and query and not on prior knowledge.
 
-Example Question:
-Can you make me lunch with ham and cheese?
-Example Response:
-INVALID
-
-Example Question:
-Why is the sky blue?
-Example Response:
-INVALID
-
-Example Question:
-Can you help configure my cluster to automatically scale?
-Example Response:
-VALID
-
-Question:
+TASK:
+{task}
+QUERY:
 {query}
+
+Please combine the information from the task and query into a single, new task.
+
 Response:
 """
         )
 
-        self.logger.info(conversation + " Validatin query")
+        self.logger.info(conversation + " Rephrasing task and query")
         self.logger.info(conversation + " usng model: " + model)
 
-        bare_llm = get_watsonx_predictor(model=model, min_new_tokens=1, max_new_tokens=2)
+        bare_llm = get_watsonx_predictor(model=model, min_new_tokens=5)
         llm_chain = LLMChain(llm=bare_llm, prompt=prompt_instructions, verbose=verbose)
 
-        task_query = prompt_instructions.format(query=query)
+        task_query = prompt_instructions.format(task=task, query=original_query)
 
         self.logger.info(conversation + " task query: " + task_query)
 
-        response = llm_chain(inputs={"query": query})
+        response = llm_chain(inputs={"task": task, "query": original_query})
 
         self.logger.info(conversation + " response: " + str(response))
         return response['text']
 
 if __name__ == "__main__":
+    """to execute, from the repo root, use python -m modules.task_rephraser.py"""
     import argparse
 
     parser = argparse.ArgumentParser(
-        description="Validate whether or not a question is about Kubernetes and OpenShift"
+        description="Rephrase a combination of a task and a user query into a single request"
     )
     parser.add_argument(
         "-c",
@@ -106,6 +97,13 @@ if __name__ == "__main__":
         default="1234",
         type=str,
         help="A short identifier for the conversation",
+    )
+    parser.add_argument(
+        "-t",
+        "--task",
+        default="1. Make a sandwich with the defined ingredients",
+        type=str,
+        help="A task statement",
     )
     parser.add_argument(
         "-q",
@@ -126,11 +124,11 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
-    question_validator = QuestionValidator()
-    question_validator.validate_question(
+    task_rephraser = TaskRephraser()
+    task_rephraser.rephrase_task(
         args.conversation_id,
+        args.task,
         args.query,
         model=args.model,
         verbose=args.verbose,
     )
-
